BEE/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.3.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
msal==1.17.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
